0428 update

+++ order-service/src/main/java/com/example/orderservice/order/domain/DeliveryAddressRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Mono;
 
 @Repository
-public interface DeliveryAddressRepository extends ReactiveCrudRepository<DeliveryAddress, Integer> {
+public interface DeliveryAddressRepository extends ReactiveCrudRepository<DeliveryAddress, Long> {
     Mono<DeliveryAddress> findByMerchantUid(String merchantUid);
 }
